lambda/requirements.txt
Pillow
boto3
requests
